Create ArduinoMiniAmmeter.ino
First upload of the ArduinoMiniAmmeter code.
Tested on an ATmega8A running on breadboard with an I2C LCD hooked up to it. Running well off of 5V from a USB buck.
Still a bit of noise present...
There'll also be an offset if the reading doesn't show zero even after calibration. Just subtract/add it from/to the final reading.